music_bot_server/lofi_music/requirements.txt
aiohttp==3.9.5
discord==2.3.2
discord.py==2.3.2
python-dotenv==1.0.1
yt-dlp==2024.4.9
